+++ projectroles/plugins.py
@@ -27,14 +27,36 @@ class ProjectAppPluginPoint(PluginPoint):
     urls = []
 
     #: Project settings definition
-    # TODO: Define project specific settings in your app plugin, example below
-    project_settings = {
-        'example_setting': {
-            'type': 'STRING',  # STRING/INTEGER/BOOLEAN
-            'default': 'example',
-            'description': 'Example setting',  # Optional
-        }
-    }
+    #:
+    #: For example ::
+    #:
+    #:     project_settings = {
+    #:         'example_setting': {
+    #:             'type': 'STRING',  # STRING/INTEGER/BOOLEAN
+    #:             'default': 'example',
+    #:             'label': 'Example project setting',  # Optional, defaults to name/key
+    #:             'placeholder': 'Enter example setting here',  # Optional
+    #:             'description': 'Example project setting',  # Optional
+    #:         }
+    #:     }
+    # TODO: Define project specific settings in your app plugin, example above
+    project_settings = {}
+
+    #: User settings (NOT project specific)
+    #:
+    #: For example ::
+    #:
+    #:     user_settings = {
+    #:         'example_setting': {
+    #:             'type': 'STRING',  # STRING/INTEGER/BOOLEAN
+    #:             'default': 'example',
+    #:             'label': 'Example user setting label',  # Optional, defaults to name/key
+    #:             'placeholder': 'Enter example setting here',  # Optional
+    #:             'description': 'Description for example user setting',  # Optional
+    #:         }
+    #:     }
+    # TODO: Define user specific settings in your app plugin, example above
+    user_settings = {}
 
     #: FontAwesome icon ID string
     # TODO: Implement this in your app plugin