+++ userprofile/urls.py
@@ -5,5 +5,10 @@ from . import views
 app_name = 'userprofile'
 
 urlpatterns = [
-    url(regex=r'^profile$', view=views.UserDetailView.as_view(), name='detail')
+    url(regex=r'^profile$', view=views.UserDetailView.as_view(), name='detail'),
+    url(
+        regex=r'^profile/settings/update$',
+        view=views.UserSettingUpdateView.as_view(),
+        name='update-settings',
+    ),
 ]